requirements.txt
pipenv
django
djangorestframework
django-cors-headers
pymysql
pillow